Second checkin with changes to mongodb
Second set of changes to get mongodb version working. All 38 tests pass
in fakeDB and mongoDB modes. We need to run the batch file that sets up
test dbs before running the test in mongodb
- Added CreateDBHandleAsync to DBCRUD interface
- changed psdb module to cache the dbHandle to a series ( in
seriesInfoMap) so we don't create a new handle every time a new token is
requested.
- updating the dust-full.min.js version to v2.5.0 which is compatible
with v0.5.1 of the dust npm module