First checkin with psMEAN client
Changed the gruntfile to copy the indexMEAN.html to
debug/public/index.html so now the server serves the MEAN client instead
of the old client. Updated the package.json to add the postinstall step
of calling bower install. Added bower.json and .bowerrc to make sure the
bower_components are placed under public/lib