Support encrypted files and writing files
Added support for reading encrypted data files, and for writing both
unencrypted and encrypted data files. The written files are generated
"from scratch" -- from the internal parsed data structures. So the
exercise of roundripping encrypted and decrypted versions of the file is
a solid verification of all aspects of reading, parsing and decrypting
data. No known bugs in this though I've not done extensive testing.