Added support for MachineName to TeamName mapping.
Implemented support for reading team name (and ID) from a csv file
(teamData.csv) that contains the team-to-machine-name mapping. So if we
populate this file with all the machine names of the laptops we are
using and their corresponding machine names then the Oracle App can
automatically pick up the name from the machine.

Added Classes:
- TeamInfo class - simple class to hold team name and id.
- CryptoHelper class (not used yet - will be used to add hash to machine
name in teamInfo.csv- maybe)
- Utils class, moved some functions like stripBlanks there.