Implementation of blocking work item queue and drive listener
- BlockingWorkQueue implements a blocking work item queue - intended for
many threads to add work that is handled by a single tread that handles
these items serially.
- NewDriveNotifier calls an event handler whenever a new drive is
detected. It does this by polling (there seems to be no dotnet way to be
notified when a new thumb drive is plugged in - the alternatives are
messy windows interop code. So instead I simply poll in a timer.