Added FileCopier, LogConsumer and re-organization of code
Reorganized the program  - it is nolonger a "sample" log processor - it
is *the* log processor. The PSDB upload code can be integrated into this
version.
This version is not working as checked in - just checking a version that
compiles before getting things to work (in some cases, getting things to
work again).