Report to user if drive is ignored because of volume label pattern mismatch
If we ignore a drive because it's volume label does not match the regex
pattern, we were just writing this fact to the debug log. Now we also
report it to the user.
Also: console CTRL-C handler now does a ReadLine not Read, and it
reports if it is continuing (vs. quitting.)