Small tweak to encrypt/decrypt - remove the (>>4) shift .
Small tweak to encrypt/decrypt - remove the (>>4) shift. That shift was
supposedly to get rid of some LSB bits because they tend to be less
random. However this is somwhat arbitrary and completely unnecessary for
our purposes because we are not trying to do anything with individual
bits - we just mod the whole thing with number of encrypt chars (the
alphabet). Updated synthesized logs to reflect this. All tested.